Add Automatic Lifts
I added both a AskForLift action for accounts to perform and a
parameter to specify how often automatic lifts should take place.
The AskForLift action doesn't work right now because I don't know
how to get the balance on individual tallies for an account. The
automatic tallies look like they are being activated, but it's hard
to tell if the lifts are actually happening or not.